app/app/requirements.txt
pymysql==1.0.2
kafka-python==2.0.2
Pillow==8.1.2
qrcode==6.1
flask-cors==3.0.10